JSON & XML Importers conforming to API v1.1 (#31)
I'm taking a liberty to merge this into `dev` based on @jlarfors approval above and @benmarsden LGTM before something else is changed in the `dev` branch this week)

Summary of changes:

* Extend JSON Importer unit test with more detailed tests

An attempt to make a more rigorous test for JSON Importer
by shotgun-type checks on some types and values in
the cty.Value structure that the Importer returns.

* Important whitespace issue in test data fixed

* Change JSON Importer unit testing philosophy

Instead of shotgun testing a few properties by a gradual
descent down the data structure, test the whole data structure
for equivalence.

* Mock up Go data structure for XML importer test file

* Add XML Importer initial implementation

There is a known issue with XML files containing lists
of length one as unlike JSON, XML does not have a syntax
for lists. This will be addressed in a separate commit.

* Reorder operations in prep for subtests

* Update fatal error handling

* Generalise XML Importer test to prep for subtests

* XML Importer subtest engulfs its private data

* Add a regression test for one element XML list

Unlike JSON with its [ ...] square brackets, XML has no
syntax for lists. Hence, the XML parser cannot tell when
some element is a standalone element or a member of a list
of unit length. This information is contained in the HCL
expression we use to convert the value returned by
the XML parser into cty.Value. A fix is in the works.

* Fix XML Importer subtests

The reporting of test failure was done incorrectly from the subtests.
When git test -v was run, the test was reported as failed, while all
of its subtests were reported as passed. Now when this command is run,
the failed subtests are reported as such.

* go modules updated

* Update Go modules

* Update Go modules

* Document the Importer API

Add comments for public methods and data structures
of the Importer.

* Separate test data and code in Importer tests

* Add a helper function for XML Importer test

* Update comments in Importer and test

Authored-by: Oliver Frolovs <ofrolovs@verifa.io>