Initial implementation of the new store
The new store uses a graphql server and writes directly to the persistent storage.
This removes the need for an interim database that was using memdb.

The graphql schema is based on the schema defined in bubbly, currently created using
core.Tables.

This work includes the initial docker and integration setup to test with a database.