Refactored the parser MASSIVELY and added validation for resource inputs
With the new way that resources can handle parsing themselves by depending
on the parser, and the fact that the parser does not need to resolve any
variables (it gets everything as input), pretty much the entire parser
could be deleted.

There is one feature in the gohcl pacakge that we need, which does not
exist and maybe never will, which is supporting cty.Types. Thus, the repo
has been forked and updated and we are using that version now via a go
mod replace.