Initial version of DataRef support
This refactors the store package. The resolver mainly becomes the new dataref.go file
which does not actually do any resolving, as the resolved values are based on some
values being saved first. Instead it processes the input DataBlocks and splits it up
ready for the Save() method to then store (and in the process, resolve) and then be
able to resolve the DataRefs.