Working version of datatree
Added a cleaner way to handle the saving of core.Data in the store.
By creating a dependency tree of data nodes that need to be solved in order, handling
data references becomes much easier.
Also a nice trick to add a traverse function with callback makes writing providers easier.