Major fixes to store migration
Wrote a more comprehensive test suite for testing the migrations.
Both the schema updates/changes which are handled by the store, and also the apply-ing of
those by the postgres provider.

Also added a test package, which I originally built in bubbly-auth and thought it could
be very useful in bubbly too. Would be good to standardise on this for the use of dockertest+postgres